refactor(subs): restructure service files
